pyproject.toml
[tool.poetry]
name = "quote-project"
version = "1.0"
description = "Simple blog, quote store"
authors = ["dima cheban <dimacheban23@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.6"
pymongo = "^4.11"
psycopg2-binary = "^2.9.10"
mongoengine = "^0.29.1"
pydantic = "^2.10.6"
pydantic-settings = "^2.7.1"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
